# api-service/src/main/gen/api/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@chokma.com
    name: Chokma Support
  description: Document extraction and processing API
  title: Chokma API
  version: 1.0.0
servers:
- description: Production
  url: https://api.chokma.com
- description: Local development
  url: http://localhost:8080
...
paths:
...
  /api/extraction/files/upload:
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/_api_extraction_files_upload_post_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
          description: File uploaded successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Upload file to Third party
      tags:
      - Extraction
      x-content-type: multipart/form-data
      x-accepts:
      - application/json
  /api/extraction/files:
    get:
      parameters:
      - explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          format: int32
          maximum: 100
          minimum: 1
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
          description: Files listed successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: List files
      tags:
      - Extraction
      x-accepts:
      - application/json
  /api/extraction/files/{fileId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: fileId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: File deleted successfully
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Delete file
      tags:
      - Extraction
      x-accepts:
      - application/json
  /api/extraction/files/{fileId}/metadata:
    get:
      parameters:
      - explode: false
        in: path
        name: fileId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileMetadata'
          description: File metadata retrieved
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Get file metadata
      tags:
      - Extraction
      x-accepts:
      - application/json
  /api/extraction/extract:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtractionRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExtractionResponse'
          description: Extraction job created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Create extraction job
      tags:
      - Extraction
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/extraction/parse:
    get:
      parameters:
      - explode: true
        in: query
        name: cursor
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseListResponse'
          description: Parse jobs listed
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: List parse jobs
      tags:
      - Extraction
      x-accepts:
      - application/json
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParseRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseResponse'
          description: Parse job created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Create parse job
      tags:
      - Extraction
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/extraction/parse/{parseId}:
    delete:
      parameters:
      - explode: false
        in: path
        name: parseId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseDeleteResponse'
          description: Parse deleted
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Delete parse job
      tags:
      - Extraction
      x-accepts:
      - application/json
    get:
      parameters:
      - explode: false
        in: path
        name: parseId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParseStatusResponse'
          description: Parse status retrieved
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Resource not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Get parse status
      tags:
      - Extraction
      x-accepts:
      - application/json
  /api/extraction/read:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadResponse'
          description: Document read initiated
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Read document
      tags:
      - Extraction
      x-content-type: application/json
      x-accepts:
      - application/json
  /api/extraction/classify:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassifyRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassifyResponse'
          description: Classification completed
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad request - invalid input
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized - authentication required
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal server error
      security:
      - bearerAuth: []
      summary: Classify document pages
      tags:
      - Extraction
      x-content-type: application/json
      x-accepts:
      - application/json
...
components:
...
  schemas:
...
    ErrorResponse:
      example:
        traceId: traceId
        code: code
        details:
          key: ""
        message: message
        timestamp: 2000-01-23T04:56:07.000+00:00
      properties:
        message:
          description: Human-readable error message
          type: string
        code:
          description: Machine-readable error code
          type: string
        timestamp:
          format: date-time
          type: string
        traceId:
          description: Request trace ID for debugging
          type: string
        details:
          additionalProperties: true
          description: Additional error details
          type: object
      required:
      - code
      - message
      type: object
...
    FileUploadResponse:
      example:
        file_id: file_id
        created_at: 2024-01-05T16:15:50Z
      properties:
        file_id:
          description: Third party file ID
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - created_at
      - file_id
      type: object
    FileMetadata:
      example:
        checksum_sha256: checksum_sha256
        mime_type: application/pdf
        file_name: file_name
        file_id: file_id
        created_at: created_at
        file_size: 0
        labels:
          key: labels
      properties:
        file_id:
          type: string
        mime_type:
          example: application/pdf
          type: string
        file_size:
          format: int64
          type: integer
        checksum_sha256:
          type: string
        created_at:
          type: string
        file_name:
          type: string
        labels:
          additionalProperties:
            type: string
          type: object
      required:
      - created_at
      - file_id
      - file_name
      - file_size
      - mime_type
      type: object
    FileListResponse:
      example:
        prev_cursor: prev_cursor
        next_cursor: next_cursor
        has_more: true
        items:
        - checksum_sha256: checksum_sha256
          mime_type: application/pdf
          file_name: file_name
          file_id: file_id
          created_at: created_at
          file_size: 0
          labels:
            key: labels
        - checksum_sha256: checksum_sha256
          mime_type: application/pdf
          file_name: file_name
          file_id: file_id
          created_at: created_at
          file_size: 0
          labels:
            key: labels
      properties:
        items:
          items:
            $ref: '#/components/schemas/FileMetadata'
          type: array
        has_more:
          type: boolean
        next_cursor:
          type: string
        prev_cursor:
          type: string
      required:
      - has_more
      - items
      type: object
    StructuredExtractionOption:
      example:
        model_provider: third party
        json_schema:
          key: ""
        partition_strategy: none
        schema_name: schema_name
        skip_ocr: false
        page_classes:
        - page_classes
        - page_classes
        prompt: prompt
        provide_citations: true
      properties:
        schema_name:
          type: string
        json_schema:
          additionalProperties: true
          description: JSON Schema for extraction
          type: object
        skip_ocr:
          default: false
          type: boolean
        prompt:
          type: string
        model_provider:
          default: third party
          type: string
        partition_strategy:
          default: none
          enum:
          - none
          - page
          - section
          type: string
        page_classes:
          items:
            type: string
          type: array
        provide_citations:
          default: true
          type: boolean
      required:
      - schema_name
      type: object
...
    EnrichmentOptions:
      example:
        figure_summarization: false
        figure_summarization_prompt: figure_summarization_prompt
        include_full_page_image: false
        table_summarization: false
        table_summarization_prompt: table_summarization_prompt
      properties:
        table_summarization:
          default: false
          type: boolean
        table_summarization_prompt:
          type: string
        figure_summarization:
          default: false
          type: boolean
        figure_summarization_prompt:
          type: string
        include_full_page_image:
          default: false
          type: boolean
      type: object
    PageClassification:
      example:
        name: name
        description: description
      properties:
        name:
          type: string
        description:
          type: string
      required:
      - name
      type: object
    ExtractionRequest:
      example:
        page_range: 1-5
        structured_extraction_options:
        - model_provider: third party
          json_schema:
            key: ""
          partition_strategy: none
          schema_name: schema_name
          skip_ocr: false
          page_classes:
          - page_classes
          - page_classes
          prompt: prompt
          provide_citations: true
        - model_provider: third party
          json_schema:
            key: ""
          partition_strategy: none
          schema_name: schema_name
          skip_ocr: false
          page_classes:
          - page_classes
          - page_classes
          prompt: prompt
          provide_citations: true
        mime_type: mime_type
        file_name: file_name
        file_id: file_id
        labels:
          key: labels
      properties:
        file_id:
          type: string
        page_range:
          example: 1-5
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        structured_extraction_options:
          items:
            $ref: '#/components/schemas/StructuredExtractionOption'
          type: array
        labels:
          additionalProperties:
            type: string
          type: object
      required:
      - file_id
      type: object
    ExtractionResponse:
      example:
        created_at: created_at
        parse_id: parse_id
      properties:
        parse_id:
          description: Third party parse/job ID
          type: string
        created_at:
          type: string
      required:
      - created_at
      - parse_id
      type: object
    ParseRequest:
      example:
        enrichment_options:
          figure_summarization: false
          figure_summarization_prompt: figure_summarization_prompt
          include_full_page_image: false
          table_summarization: false
          table_summarization_prompt: table_summarization_prompt
        page_range: page_range
        structured_extraction_options:
        - model_provider: third party
          json_schema:
            key: ""
          partition_strategy: none
          schema_name: schema_name
          skip_ocr: false
          page_classes:
          - page_classes
          - page_classes
          prompt: prompt
          provide_citations: true
        - model_provider: third party
          json_schema:
            key: ""
          partition_strategy: none
          schema_name: schema_name
          skip_ocr: false
          page_classes:
          - page_classes
          - page_classes
          prompt: prompt
          provide_citations: true
        mime_type: mime_type
        file_name: file_name
        file_id: file_id
        parsing_options:
          chunking_strategy: none
          table_parsing_format: tsr
          skew_detection: false
          table_output_mode: html
          disable_layout_detection: false
          cross_page_header_detection: false
          signature_detection: false
          remove_strikethrough_lines: false
          barcode_detection: false
          ignore_sections:
          - ignore_sections
          - ignore_sections
          include_images: false
          ocr_model: model01
        page_classifications:
        - name: name
          description: description
        - name: name
          description: description
        labels:
          key: labels
      properties:
        file_id:
          type: string
        page_range:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        parsing_options:
          $ref: '#/components/schemas/ParsingOptions'
        structured_extraction_options:
          items:
            $ref: '#/components/schemas/StructuredExtractionOption'
          type: array
        page_classifications:
          items:
            $ref: '#/components/schemas/PageClassification'
          type: array
        enrichment_options:
          $ref: '#/components/schemas/EnrichmentOptions'
        labels:
          additionalProperties:
            type: string
          type: object
      required:
      - file_id
      type: object
    ParseResponse:
      example:
        created_at: 2024-01-05T16:15:50Z
        parse_id: parse_id
      properties:
        parse_id:
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - created_at
      - parse_id
      type: object
    ParseStatusResponse:
      example:
        parsed_pages_count: 0
        finished_at: 2024-01-05T16:15:50Z
        chunks:
        - key: ""
        - key: ""
        dataset_id: dataset_id
        usage:
          key: ""
        created_at: 2024-01-05T16:15:50Z
        parse_id: parse_id
        total_pages: 6
        page_classes:
          key: ""
        error: error
        message_update: message_update
        labels:
          key: ""
        pdf_base64: pdf_base64
        pages:
        - key: ""
        - key: ""
        tasks_completed_count: 1
        options:
          key: ""
        structured_data:
          key: ""
        status: pending
        tasks_total_count: 5
      properties:
        parse_id:
          type: string
        status:
          enum:
          - pending
          - processing
          - completed
          - failed
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
        dataset_id:
          type: string
        parsed_pages_count:
          format: int32
          type: integer
        total_pages:
          format: int32
          type: integer
        error:
          type: string
        pages:
          items:
            additionalProperties: true
            type: object
          type: array
        chunks:
          items:
            additionalProperties: true
            type: object
          type: array
        structured_data:
          additionalProperties: true
          type: object
        page_classes:
          additionalProperties: true
          type: object
        pdf_base64:
          format: byte
          type: string
        tasks_completed_count:
          format: int32
          type: integer
        tasks_total_count:
          format: int32
          type: integer
        finished_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
        labels:
          additionalProperties: true
          type: object
        options:
          additionalProperties: true
          type: object
        usage:
          additionalProperties: true
          type: object
        message_update:
          type: string
      required:
      - created_at
      - parse_id
      - status
      type: object
    ParseListResponse:
      example:
        prev_cursor: prev_cursor
        next_cursor: next_cursor
        has_more: true
        items:
        - parsed_pages_count: 0
          finished_at: 2024-01-05T16:15:50Z
          chunks:
          - key: ""
          - key: ""
          dataset_id: dataset_id
          usage:
            key: ""
          created_at: 2024-01-05T16:15:50Z
          parse_id: parse_id
          total_pages: 6
          page_classes:
            key: ""
          error: error
          message_update: message_update
          labels:
            key: ""
          pdf_base64: pdf_base64
          pages:
          - key: ""
          - key: ""
          tasks_completed_count: 1
          options:
            key: ""
          structured_data:
            key: ""
          status: pending
          tasks_total_count: 5
        - parsed_pages_count: 0
          finished_at: 2024-01-05T16:15:50Z
          chunks:
          - key: ""
          - key: ""
          dataset_id: dataset_id
          usage:
            key: ""
          created_at: 2024-01-05T16:15:50Z
          parse_id: parse_id
          total_pages: 6
          page_classes:
            key: ""
          error: error
          message_update: message_update
          labels:
            key: ""
          pdf_base64: pdf_base64
          pages:
          - key: ""
          - key: ""
          tasks_completed_count: 1
          options:
            key: ""
          structured_data:
            key: ""
          status: pending
          tasks_total_count: 5
      properties:
        items:
          items:
            $ref: '#/components/schemas/ParseStatusResponse'
          type: array
        has_more:
          type: boolean
        next_cursor:
          type: string
        prev_cursor:
          type: string
      required:
      - has_more
      - items
      type: object
    ParseDeleteResponse:
      example:
        trace_id: trace_id
        code: code
        details:
          key: ""
        message: message
        timestamp: 0
      properties:
        message:
          type: string
        code:
          type: string
        timestamp:
          format: int64
          type: integer
        trace_id:
          type: string
        details:
          additionalProperties: true
          type: object
      required:
      - code
      - message
      - timestamp
      type: object
    ReadRequest:
      example:
        enrichment_options:
          figure_summarization: false
          figure_summarization_prompt: figure_summarization_prompt
          include_full_page_image: false
          table_summarization: false
          table_summarization_prompt: table_summarization_prompt
        page_range: page_range
        mime_type: mime_type
        file_name: file_name
        file_id: file_id
        parsing_options:
          chunking_strategy: none
          table_parsing_format: tsr
          skew_detection: false
          table_output_mode: html
          disable_layout_detection: false
          cross_page_header_detection: false
          signature_detection: false
          remove_strikethrough_lines: false
          barcode_detection: false
          ignore_sections:
          - ignore_sections
          - ignore_sections
          include_images: false
          ocr_model: model01
        labels:
          key: labels
      properties:
        file_id:
          type: string
        page_range:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        parsing_options:
          $ref: '#/components/schemas/ParsingOptions'
        enrichment_options:
          $ref: '#/components/schemas/EnrichmentOptions'
        labels:
          additionalProperties:
            type: string
          type: object
      required:
      - file_id
      type: object
    ReadResponse:
      example:
        created_at: 2024-01-05T16:15:50Z
        parse_id: parse_id
      properties:
        parse_id:
          type: string
        created_at:
          example: 2024-01-05T16:15:50Z
          format: date-time
          type: string
      required:
      - created_at
      - parse_id
      type: object
    ClassifyRequest:
      example:
        page_range: page_range
        mime_type: mime_type
        file_name: file_name
        file_id: file_id
        page_classifications:
        - name: name
          description: description
        - name: name
          description: description
        labels:
          key: labels
      properties:
        file_id:
          type: string
        page_range:
          type: string
        file_name:
          type: string
        mime_type:
          type: string
        page_classifications:
          items:
            $ref: '#/components/schemas/PageClassification'
          type: array
        labels:
          additionalProperties:
            type: string
          type: object
      required:
      - file_id
      type: object
    ClassifyResponse:
      example:
        result:
          key: ""
        status: status
      properties:
        status:
          type: string
        result:
          additionalProperties: true
          type: object
      required:
      - status
      type: object
...
    _api_extraction_files_upload_post_request:
      properties:
        file:
          format: binary
          type: string
        labels:
          description: JSON string of labels
          type: string
      required:
      - file
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token authentication
      scheme: bearer
      type: http